--- test/ng-build/ng5/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "ng5",
-  "version": "0.0.0",
-  "license": "MIT",
-  "scripts": {
-    "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
-  },
-  "private": true,
-  "dependencies": {
-    "@angular/animations": "5.0.0",
-    "@angular/common": "5.0.0",
-    "@angular/compiler": "5.0.0",
-    "@angular/core": "5.0.0",
-    "@angular/forms": "5.0.0",
-    "@angular/http": "5.0.0",
-    "@angular/platform-browser": "5.0.0",
-    "@angular/platform-browser-dynamic": "5.0.0",
-    "@angular/router": "5.0.0",
-    "angularfire2": "file:../../../dist/packages-dist",
-    "core-js": "^2.4.1",
-    "rxjs": "^5.5.0",
-    "zone.js": "^0.8.14"
-  },
-  "devDependencies": {
-    "@angular/cli": "1.5",
-    "@angular/compiler-cli": "5.0.0",
-    "@angular/language-service": "5.0.0",
-    "typescript": "~2.4.2"
-  }
-}
