--- test/ng-build/ng5/src/app/app.module.ts
@@ -1,25 +1,0 @@
-import { BrowserModule } from '@angular/platform-browser';
-import { NgModule } from '@angular/core';
-import { FirebaseAppConfig } from 'angularfire2';
-import { AngularFireDatabaseModule } from 'angularfire2/database';
-import { AngularFireAuthModule } from 'angularfire2/auth';
-import { AngularFirestoreModule } from 'angularfire2/firestore';
-
-import { AppComponent } from './app.component';
-
-@NgModule({
-  declarations: [
-    AppComponent
-  ],
-  imports: [
-    BrowserModule,
-    AngularFireAuthModule,
-    AngularFireDatabaseModule,
-    AngularFirestoreModule
-  ],
-  providers: [
-    { provide: FirebaseAppConfig, useValue: {} }
-  ],
-  bootstrap: [AppComponent]
-})
-export class AppModule { }
