+++ test/ng-build/ng6/.angular-cli.json
