+++ test/ng-build/ng6/package.json
@@ -1,0 +1,37 @@
+{
+  "name": "ng5",
+  "version": "0.0.0",
+  "license": "MIT",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "build:prod": "ng build --prod",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e"
+  },
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "6.0.1",
+    "@angular/common": "6.0.1",
+    "@angular/compiler": "6.0.1",
+    "@angular/core": "6.0.1",
+    "@angular/forms": "6.0.1",
+    "@angular/http": "6.0.1",
+    "@angular/platform-browser": "6.0.1",
+    "@angular/platform-browser-dynamic": "6.0.1",
+    "@angular/router": "6.0.1",
+    "angularfire2": "file:../../../dist/packages-dist",
+    "core-js": "^2.4.1",
+    "rxjs": "^6.0.0",
+    "zone.js": "^0.8.14"
+  },
+  "devDependencies": {
+    "@angular/cli": "6.0.1",
+    "@angular/compiler-cli": "6.0.1",
+    "@angular/language-service": "6.0.1",
+    "typescript": ">=2.7.2 <2.8.0",
+    "@angular-devkit/build-angular": "~0.6.1"
+  }
+}
