+++ test/ng-build/ng6/src/app/app.component.ts
