+++ test/ng-build/ng6/src/environments/environment.prod.ts
