+++ test/ng-build/ng6/src/index.html
