+++ test/ng-build/ng6/src/main.ts
