+++ test/ng-build/ng6/src/tsconfig.app.json
