+++ test/ng-build/ng6/src/typings.d.ts
